Simplify hook implementation.
	modified:   pm/HOWTO.hooks
	modified:   pm/functions.in
	modified:   pm/sleep.d/Makefile.am
	deleted:    pm/sleep.d/zzz

* Merge the functionality of find_hooks and run_hooks.  This results in
a new run_hooks function which is much simpler than the old one and is
much more resistant to breaking due to odd filenames.
* Actually implements the functionality mentioned in README where you
can create a nonexecutable file in /etc/pm/(sleep|power).d that will
prevent the copy in /usr/lib/pm-utils/(sleep|power).d from running.
* Undoes the "nonzero exit code means do not run on resume" and
auto-reverse pseudofeatures I added in the POSIX series. If anyone likes
them, they can be added back at the cost of either passing a function to
run_hooks or maintaining a global variable.
* Merges the functionality that was in zzz back into pm_main. Although
zzz was a slightly funny name, upon further reflection it was a bad idea
and makes it harder to implement some partitioning of functionality that
will be needed if/when the hooks are split into their own project.
* Updates the documentation to reflect these changes.