Much simpler autogen.sh for our needs
The current autogen.sh script had way too much baggage for what we use
in pm-utils. My guess is it was copied from a GNOME package.

Anyway, this slims down autogen.sh to just use autoreconf, which will do
the right thing. This autogen.sh is a virtual copy of the "official"
script used in Xorg.