Split functions.in into functions.in and pm-functions.in
pm-functions.in contains the functionality that should only be used by the
pm-utils frontends (pm-action, pm-is-supported, and pm-powersave).

functions.in contains code that is shared between the hooks and the pm-utils
frontends.

The reason behind splitting this out is to reduce namespace pollution in
the hooks and to eventaully make it easier to split the hooks into their
own package.