Richard raised concerns about hard drive spindown.
So the harddrive hook no longer tries to spin the drive down pending
some sort of consensus on the right thing to do to make the hard drive
save power.

In the mean time, though, just flat-out disable write caching on the
hard drive.  I know that there are barriers and flushes that the ATA layer
uses to ensure that data really is written to the disk, but hard drive cache
isn't battery backed.