Drop 50ntpd
ntpd operates normally over a sleep/wake cycle. Stoppping and
restarting it in a sleep hook is too adhoc. If it is essential
that ntpd be stopped and restarted, then having NetworkManager
do the job when interfaces go offline and come back online is
a better way to do things.