Make kernel modesetting detection a little smarter.
It turns out that the method we were using to detect kernel modesetting
support was not very accurate.  It turns out there is no bulletproof way
for us to tell if KMS is being used, but Michael Biebl found a way that
sucks less than out current method.  Hopefully the framebuffer drivers
will grow a flag in sysfs somewhere that tells us that KMS is in use.