Don't err out in maybe_deallocvt if there is no console state
kpowersave pops up a window telling that an error "1" has occurred when
suspending to ram. The error "1" message is annoying and lacks meaning
to the user.

It's caused by maybe_deallocvt returning the ret_val of state_exists
when the tested (here: console) file isn't found. This absence of this file
seems not to do any harm therefore the escalation of the ret_val 1 is
deprecated.

Solution: just tell maybe_deallocvt to return 0 in the above case.

http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=554959